Move type icons into templates use generic tx metadata
This moves around the emoji to be in a template rather than the code,
and also adds icons for registration and contributions.

It also reuses the existing tx SN metadata code (which I moved out into
a separate function), replacing the separate logic that was only be used
for the added emoji.